05 - Finalizando - Desafio

+++ nosso_primeiro_projeto/lib/components/task.dart
@@ -9,7 +9,8 @@ class Task extends StatefulWidget {
   final String foto;
   final int dificuldade;
 
-  Task(this.nome, this.foto, this.dificuldade, {super.key});
+  Task(this.nome, this.foto, this.dificuldade, this.nivel, this.maestryLevel,
+      {super.key});
 
   int nivel = 0;
   int maestryLevel = 0;
@@ -87,12 +88,36 @@ class _TaskState extends State<Task> {
                     width: 52,
                     child: ElevatedButton(
                       onLongPress: () {
-                        TaskDao().delete(widget.nome);
+                        showDialog(
+                          context: context,
+                          builder: (BuildContext context) {
+                            return AlertDialog(
+                              title: Text('Confirmação'),
+                              content: Text('Confirma a exclusão do registro?'),
+                              actions: [
+                                TextButton(
+                                  child: Text('Sim'),
+                                  onPressed: () {
+                                    TaskDao().delete(widget.nome);
+                                    Navigator.of(context).pop();
+                                  },
+                                ),
+                                TextButton(
+                                  child: Text('Não'),
+                                  onPressed: () {
+                                    Navigator.of(context).pop();
+                                  },
+                                ),
+                              ],
+                            );
+                          },
+                        );
                       },
                       onPressed: () {
                         setState(() {
                           widget.nivel++;
                           defineMaestryLevel();
+                          saveTask();
                         });
                         //print(nivel);
                       },
@@ -151,5 +176,10 @@ class _TaskState extends State<Task> {
       widget.maestryLevel++;
       widget.nivel = 0;
     }
+  }
+
+  void saveTask() {
+    TaskDao().save(Task(widget.nome, widget.foto, widget.dificuldade,
+        widget.nivel, widget.maestryLevel));
   }
 }
